Spark/requirements.txt
numpy
pandas
requests
Pillow
plotly
streamlit
confluent-kafka==1.7.0
python-dateutil==2.8.2